+++ dynatrace-appmon/src/main/java/de/tsystems/mms/apm/performancesignature/dynatrace/rest/json/ApiClient.java
@@ -258,7 +258,7 @@ public class ApiClient {
                     throw new ApiException(response.message(), e, response.code(), response.headers().toMultimap());
                 }
             }
-            throw new ApiException(response.message(), response.code(), response.headers().toMultimap(), respBody);
+            throw new ApiException(respBody, response.code(), response.headers().toMultimap(), respBody);
         }
     }